+++ exporters/xrayudp/exporter_test.go
@@ -1,7 +1,7 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package otlptraceudp
+package xrayudp
 
 import (
 	"context"
@@ -12,12 +12,7 @@ import (
 
 func TestOtlpUdpExporter(t *testing.T) {
 	t.Run("Create Exporter", func(t *testing.T) {
-		_, err := New(context.TODO(), WithEndpoint("1.2.3.4:9876"), WithSignalPrefix("E3"))
-		assert.NoError(t, err)
-	})
-
-	t.Run("Create Unstarted Exporter", func(t *testing.T) {
-		_, err := NewUnstarted(WithEndpoint("1.2.3.4:9876"), WithSignalPrefix("E3"))
+		_, err := NewSpanExporter(context.TODO(), WithEndpoint("1.2.3.4:9876"), WithSignalPrefix("E3"))
 		assert.NoError(t, err)
 	})
 }
