+++ exporters/xrayudp/udp_exporter_test.go
@@ -1,7 +1,7 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package otlptraceudp
+package xrayudp
 
 import (
 	"encoding/base64"
